carro/requirements.txt
paho-mqtt
requests
geopy
paho-mqtt
